Moved @tryghost/members-events-service code into Ghost core
fix https://linear.app/ghost/issue/ENG-2066/members-events-service

- this commit migrates the code and test for the members events service
  into Ghost core so we can keep all related code together and reduce
  overhead
- as a result of this change, we can remove the package and associated
  references
