+++ ghost/core/package.json
@@ -122,7 +122,6 @@
     "@tryghost/member-events": "0.0.0",
     "@tryghost/members-api": "0.0.0",
     "@tryghost/members-csv": "0.0.0",
-    "@tryghost/members-events-service": "0.0.0",
     "@tryghost/members-importer": "0.0.0",
     "@tryghost/members-offers": "0.0.0",
     "@tryghost/members-ssr": "0.0.0",
