--- ghost/members-events-service/lib/index.js
@@ -1,5 +1,0 @@
-module.exports = {
-    LastSeenAtUpdater: require('./LastSeenAtUpdater'),
-    LastSeenAtCache: require('./LastSeenAtCache'),
-    EventStorage: require('./EventStorage')
-};
